jestify addtl info tests
